Add Product entity support with mapping and service logic improvements
- Added `Products` DbSet to `ComputergyDbContext`.
- Changed `Price` type from `decimal` to `float` in `ProductDTO`.
- Corrected method signature for `GetOneProduct` to allow returning `null`.
- Created `ProductMappings` class to map `Product` entities to `ProductDTO`.
- Implemented basic `GetOneProduct` logic in `ProductsService` to fetch a product by ID.

+++ ComputergyAPI/DTOs/Products/ProductDTO.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         public string? ImageUrl { get; set; }
         public string Category { get; set; }
         public string Brand { get; set; }
